// LDC.Atlas.WebApp/Client/app/shared/entities/additional-cost.entity.ts
export class AdditionalCost {
    cashAdditionalCostId: number;
    costDirectionId: number;
    costTypeCode: string;
    nominalAccountCode: string;
    accountLineType: string;
    currencyCode: string;
    amount: number;
    narrative: string;
    accountId: number;
    accountCode: string;
    clientAccount: number;
    counterpartyCode: string;
    nominalAccountNumber: string;
    customerVendor: string;
    clientAccountMandatory: boolean;
}