// LDC.Atlas.WebApp/Client/app/shared/entities/nominal-account.entity.ts
export interface NominalAccount {
    accountNumber: string;
    nominalAccountNumberFormated: string;
    detailedDescription: string;
    shortDescription: string;
    nominalAccountId: number;
    detailAccountTitle: string;
    bankAccount: boolean;
    alternativeDescription: string;
    clientAccountMandatory: number;
    accountType: string;
    customerVendor: string;
    mainAccountTitle: string;
    displayName: string;
}